// frontend/src/api/generated/model/gitHubLink.ts
/**
 * Generated by orval v7.21.0 🍺
 * Do not edit manually.
 * Gantry Board API
 * AI Agent Orchestration Kanban Board API
 * OpenAPI spec version: 0.1.0
 */
import type { GitHubLinkLastSyncedAt } from './gitHubLinkLastSyncedAt';

export interface GitHubLink {
  created_at: string;
  id: string;
  last_synced_at?: GitHubLinkLastSyncedAt;
  project_id: string;
  repo_name: string;
  repo_owner: string;
  sync_enabled: boolean;
  updated_at: string;
}